(WIP) Merge pull requests from github reporter
Github reporter can be configured to merge pull reqeusts.

Pre-release version of Github3.py has to be used, because the latest
released version 9.4 has a bug in merge method. Furthermore the newest
merge method supports specifying exact sha to be merged, which is
desirable to ensure that the exact commit that went through the pipeline
gets merged.

Change-Id: I0c3abbcce476774a5ba8981c171382eaa4fe0abf