Install and lock uritemplate.py
This is a workaround for bug
https://github.com/sigmavirus24/github3.py/issues/634

Which is in fact between github3.py and uritemplate.py (which was
recently deprecated in favour of uritemplate). As a quick fix, adding
and locking this dependency to zuul as we currently don't have control
over the github3.py.

Revert once the bug in github3.py is fixed.

Change-Id: Ifdd88df7379caa400c9605026da8903ddcb6cd62