Run tests during RPM build
Tox is not used, because there are problems with the package
installation - pypi is not available during build phases. Therefore the
depedencies are downloaded during `make sources` and installed from pip
cache in `%build` and `%check` steps.

It would be possible to use alternative `install_command` for tox,
however, we would need to use custom tox.ini as it can't be overriden
via tox parameters.

Change-Id: I80e1f598c364d75652853dff2712810f609a0ea0